multi: Introduce SessionToken grpc parameter
This adds the SessionToken grpc return value to the FindMatchesResponse
structure and to the GenerateTicketRequest, FundTicketRequest and
FundSplitRequest messages.

This is used as a unique key to the participant's session, offering a
way to verify whether the further requests to complete the session are
coming from the appropriate participant.

This is currently implemented as a random, 32 byte token.