matcher: Test for correct stage participation
This adds a CurrentStage property to both Session and
SessionParticipant, such that the matcher can verify whether the request
sent by the buyer corresponds to the correct expected request for the
current stage of the session.

This adds protection against double participation attempt (trying to
send the same request twice) and rogue participation attempts (trying to
send data for the wrong stage of the session).