buyer: Treat failed DNS SRV queries as warnings
Instead of treating DNS SRV lookups as failures, treat them as warnings,
report them to the user and still try to connect to the original matcher
host.

This is needed because detection of the "no such host" error is fragile
and platform-dependent, so instead of simply failing to connect and
perform the session we just continue trying with the currently
configured host and leave it up to the user to resolve problems related
to the DNS setup.