matcher: Rate-limit waiting list notifications
This changes the behavior of the waiting list watcher in the matcher so
that notifications are rate limited to once every five seconds.

This solves a potential DoS vector by someone spamming connections. It
also fixes a possible race condition that caused blocking notifications
at the start of new sessions.