multi: Change SecretNumber to []byte
This changes the SecretNumber type from an uint64 to a slice of bytes,
such that it can be correctly represented even in architectures that
lack 64 bit integers (eg: wasm/js).

Given the change in the definition of the matcher api, this requires a
protocol bump to version 5.

Notice that given how this change was implemented, the matcher accepts
secret numbers of any length, so it's the client's reposibility to
generate a byte slice with sufficient entropy (currently hard coded as 8
- the same as the older uint64 implementation).

While the tests were fixed to support the new type, the specific checks
performed for lottery commitment verification did not need to change,
since the older values stored in uint64 could simply be converted into
little endian byte slices.