matcher: Sanity check sigscript on fundSplitTx
This adds a sanity check of the length of a participant's provided
signature script for split transactions.

This is needed to track down a specific error path where an empty (or
too big and therefore invalid) signature script is sent from a
participant, causing the session to never be considered as finished.