splitticket: Only request dcrdata tx once
This is a small fix when dcrdata is used for fetching utxo data.
Previously the same transaction could be requested twice, due to
tracking the wrong information when building the request.

While requesting this data twice can't cause duplicated elements from
being returned (due to the return value for UtxoMapOutpointsFromDcrdata
being a map indexed by the requested outpoints and hence only a single
element is returned per outpoint) this is inefficient since it wastes
bandwidth by returning duplicated tx data.